improved delete routes

+++ controllers/turmaController.js
@@ -1,5 +1,7 @@
 const TurmaModel = require('../models/Turmas.js')
 
+const keyAdmin = require('../admin/key.js')
+
 const createTurma = async (req, res) => {
     try {
         const turma = {
@@ -8,20 +10,20 @@ const createTurma = async (req, res) => {
         }
 
         if (turma.ano != "3") {
-            res.status(500).json({message: "No momento não há como utilizar outro valor a não ser '3'"})
+            res.status(500).json({ message: "No momento não há como utilizar outro valor a não ser '3'" })
             return
         }
 
         const isTurmaExistent = await TurmaModel.findOne({ identificador: turma.identificador, ano: turma.ano })
 
-        if(isTurmaExistent) {
-            res.status(500).json({message: "Essa turma já está registrada"})
+        if (isTurmaExistent) {
+            res.status(500).json({ message: "Essa turma já está registrada" })
             return
-        }        
+        }
 
         const response = await TurmaModel.create(turma)
 
-        res.status(201).json({response, message: "Turma criada com sucesso"})
+        res.status(201).json({ response, message: "Turma criada com sucesso" })
     } catch (error) {
         console.log('[CONTROLLER TURMA CREATE] Error: ' + error)
     }
@@ -43,13 +45,13 @@ const deleteTurma = async (req, res) => {
 
         const turma = await TurmaModel.findById(id)
 
-        if(!turma) {
-            res.status(404).json({message: "Turma não encontrada"})
+        if (!turma) {
+            res.status(404).json({ message: "Turma não encontrada" })
         }
 
-       const deletedTurma = await TurmaModel.findByIdAndDelete(id)
+        const deletedTurma = await TurmaModel.findByIdAndDelete(id)
 
-       res.status(200).json({deletedTurma, message: "Turma excluída com sucesso"})
+        res.status(200).json({ deletedTurma, message: "Turma excluída com sucesso" })
     } catch (error) {
         console.log('[CONTROLLER TURMA DELETE] Error: ' + error)
     }
@@ -57,9 +59,15 @@ const deleteTurma = async (req, res) => {
 
 const deleteAllTurma = async (req, res) => {
     try {
-       const deletedTurmas = await TurmaModel.deleteMany({})
+        const keyUrl = req.params.key
 
-       res.status(200).json({deletedTurmas, message: "Turmas excluídas com sucesso"})
+        if(keyUrl != keyAdmin) {
+            return res.status(500).json({ message: "Senha incorreta, não será possível deletar as turmas" })
+        }
+        
+        const deletedTurmas = await TurmaModel.deleteMany({})
+
+        res.status(200).json({ deletedTurmas, message: "Turmas excluídas com sucesso" })
     } catch (error) {
         console.log('[CONTROLLER TURMA DELETE ALL] Error: ' + error)
     }
